Get rid of tty->stop() code - it never worked.
This was #defined out anyways.
I spent a few weeks trying to get it to work,
so I wouldn't have to patch the kernel.
I even worked with samuel thibault on it for a while;
but it just isn't feasible.
And with a usb keyboard, it throws errors and
printk messages like crazy as the usb event queue overflows.
So let's just clear that stuff out and clean up the code a bit.