Allocate tty buffers as needed, and free them on module unload.
This is a significant restructure.
Try to allocate a buffer for console 7 when you switch to console 7
(foreground), or when console 7 has tty output.
Buffer is allocated in the vt notifier, hope that's ok.
Buffers are reset when the device is closed and reopened, but not freed.
That would be a waste; for we would almost surely reallocate them again.
Failure to allocate sends a stock message down to user space,
as though there were a fixed buffer saying "cannot allocate".
With gigs of ram, this should never happen.