Rewrite acsint driver and bridge to support unicodes.
Yes, this is a major rewrite, and I've probably broken something along the way.
Sorry about that.
The acsint device driver now captures unicodes from the tty,
and passes them down to the bridge.
These are converted back to ascii and passed to your adapter
using the functions you already know.
Thanks to encapsulation, you shouldn't have to change anything,
or at least not much of anything.
I got rid of the typedef achar, because an acsint character
is now a unicode, at least internally.
When I downshift this back to ascii, it may as well be type char.
All your string functions should now compile without warning messages.
The downshift preserves unicodes that are below 256,
that is, iso8859-1, so this should work for
many western European languages.  That will have to do for now.
I modified the todo file to describe some future international work,
but for now the first step is taken.
The driver itself speaks unicode,
and I'm sure the Linux community will smile on that when we
submit it for acceptance in the kernel.
As I say, the bridge can catch up later.
Let me know what is broken as a result of this rewrite.  Thanks.