Add comments regarding utf8 chars in macros.
This is just comments in acsbridge.h,
and also a change in the control #7 macro in acstest.cfg.
I put in some utf8 chars,
Greek begins with alpha beta gamma
verified that they pass back through the tty properly,
and then documented this fact.
So perhaps we don't have to write unicode versions of
every function - some of the functions that are naturally stream oriented,
like macros, just work as is, under utf8.