Add commands to acstest to test getsentence().
This is a complicated function, and I need a way to test it
without embeding it in a full scale adapter.
alt g now grabs the firsst line of the buffer, using getsentence(),
and prints it.
(Obviously you need another adapter to read it.)
alt w grabs the first word, using getsentence().
Use alt c to clear the buffer, then type in some new text,
then use alt g to getsentence on that line.