Move key binding functions to a new source file.
acsbridge.c was approaching 3,000 lines of code, which is a bit large
for my taste.  It can probably be broken up into smaller chunks.
One such chunk is assigning speech functions and macros to keys,
setting the pronunciation of punctuation marks, and building the
word replacement dictionary.
And of course the function acs_line_configure that ties all this together.
This is now in acsbind.c.
This is the file we will (eventually) enhance with higher unicodes,
setting the pronunciation of greek letters, math symbols,
double letter combinations, and so on.

This commit should have no effect on the interface, or any adapter,
it is just an internal restructuring of code.

I may want to pull out other chunks as well, such as the ss functions.