Open the serial port even though cdc is low.
I had to read several articles to get a handle on this one.
The linux tty ports have some unpleasant legacy behaviors
from a time when modems over phone lines were state of the art.
ttyS0 will hang on open, until the modem control line known as carrier detect,
or cdc, goes high.  Thus there is a connection, and a reason to open the port.
/dev/cau0 is for outgoing calls, and will open regardless of cdc.
So says one article,
http://en.wikipedia.org/wiki/Data_Carrier_Detect
but it doesn't work.
Other articles say cua* is depricated, and not supported.
They claim ttyS0 can behave like cua0, thus no need for cua0.
Fine, but that article didn't tell me how to do it.
Another article said to set CLOCAL in the stty control flags.
The manual page agrees; that will do the trick.
But you can't do that until the device is already open.
So it's a chicken and egg problem.
The solution is to open the device nonblock,
set CLOCAL, then use fcntl to put it back into blocking mode.
And that's what this commit does.

Some articles say CLOCAL ignores cdc, other articles say it ignores
cdc and dsr.
I'm not sure which is right, and I'm not even sure which way I
want it to be.
I think every external synth asserts dsr, so that would be a useful
signal to track.
I added the function getModemStatus to the bridge,
so we can get our hands on these signals if need be.

I don't know if this is the last word on opening serial ports,
but it seems to work now, for me, whereas it would just hang before.
I won't know for sure until I am sending index markers back and forth,
and relying on flow control etc to keep everything in sync.