Remove acs_endsentence(); it was only a stub anyways.
This is a very necessary function, but it's very English centered,
and probably doesn't belong in the bridge.
Well if I could table drive it I might still put it in the bridge,
but I can't, it's very specific software per language.
And to my surprise, I already have it in jupiter's text preprocessor.
Run jupiter ltts
and type this in
I always wanted to talk to Mr. Flintstone, and here he is!  How do you do?
This is the output.
i always wanted to talk to mister flintstone comma and here he is bang
how do you do quess
It has already put in the line break at the sentence boundary.
And it was not tricked by Mr. Flintstone.
So my adapter can read sentence by sentence.
But I don't have, integrated into this function,
an understanding that line breaks in the input might, perhaps,
turn back into spaces if we are reading continuously,
and not stopping at every line, and if noises are turned off,
so that newline and bell sounds don't have to be in sync with the text.
so there is more work to do here.
In any case, the sentence breakup function probably doesn't belong in the bridge,
and a good piece of it is already written somewhere in my
jupiter text preprocessor.  So let's leave it there for now.