Fix bug in smartreplace when rhs contains spaces.
In my dictionary I replaced sourcefile with source file,
but sourcefiles did not get similarly replaced.
I wasn't allowing for spaces in the right side.
I do now, and have documented it accordingly;
the replacement for a root word must contain letters and spaces
if smartreplace is to put the suffix back on.