Make ACS_SS_ALL practical.
This is the shift state to use when you want everything.
It wasn't practical for letters though,
because you would intercept x, as well as alt control x etc,
and you don't want to do that.
This patch makes exception for plain or shifted letters,
or alt function keys, so it is somewhat more practical,
now, to capture all flavors of a key.
But you still have to do everything yourself.
The bridge doesn't understand chords.
It can bind a speech comand to alt x, or control x, but not alt control x.