The adapter can specify the gap of time that separates new output from old.
This is the acs_obreak command.
The default is 5, or half a second.
0 turns off the feature entirely.

Also added a place holder for the acs_swoop() command,
which is not yet implemented.

Finally, the driver buffer is always 64K no matter what.
This will be ratcheted down to something smaller in user space
if the adapter wishes.
Not something I wanted really, but Samuel suggests it.

This is one of those commits that you have to rebuild everything
or it will blow up.
That's cause symbols in acsint.h changed.
Doesn't change the api though, except for the new function acs_obreak().