Allow for capture of alt or control digits and punctuations, as we do for letters.
This is backward compatible with earlier config files.
But you can now have, in your config file:
^,      (control comma)
@.     (alt period)
l@/
+^3
^@=
^]
and so on.
No, I don't let you capture plain or shifted keys on the lower 48,
as these are suppose to be ascii characters that everyone types.
But things like control apostrophe we should have,
as they are just as convenient to the home row as control j etc.