Add the new musical step function acs_scale().
This will not be accessible to you unless you have patched the kernel,
whence the step function is built into pcspkr.c.
I'm predicting some form of rapidly ascending scale
will have to replace the swoop of the newline character,
due to concerns about cpu hogging in low level kernel functions.
The adapter will use steps instead of a swoop for newline.
Too bad, I'm going to miss the swoop.
Ascending scales are just so clunky, not smooth, not as pretty.