make the bridge a shared library
This commit makes the bridge a shared library and adds support for
installing/uninstalling it to the makefile.

After this is in the tree, once the bridge library is built, you need to
run "make install" and "ldconfig" as root.

The ldconfig command should not be added to the Makefile; package
managers will take care of that once we do a release and are accepted
into distributions.