Keep pending keys in a stack of structures.
Each structure holds information about the key, including the type.
Thus a unicode can displace a keysym, can displace a keycode,
though this is not entirely implemented yet.

I also had to change the ifdef for clicks / acsint separation,
since it was not correct for modules, only for builtin.
I'm not sure which header file to include to get CONFIG_TTYCLICKS,
but I guess it's being pulled in by the includes we already have.