Use approximately the same code in pcspkr.c and locally here.
This should make it easier to move between the two;
whether we are rolling our own sounds or using the sounds in the kernel.
Only difference is cr, which is done by a swoop locally,
but has to be a step function if done in pcspkr.c,
as I don't think swoop will ever make it into the kernel proper.
step(2900,3600,10,8)