Add the keyboard half-qwerty module.
This is something Chris and I wrote 4 years ago.
I just didn't want to lose it in the shuffle.
Better to store it in git, and then it's there for anybody who might want it.

hq is not terribly compatible with acsint;
see the comments in the source under notifyer priority.