Fix the timing bug between cr sound and tty in 3.11.9.
This has been coming on for years.
The swoop that I generate for carriage return holds the cpu
for maybe a tenth of s econd, that is now an eternity.
The new tty in 3.9.11 has a separate thread for cooking,
or something, or some reason that the next echo crlf can get delayed,
if the last crlf is held up making the swoop sound.
So the swoop sound has to go.
I've been expecting this for a long time.
I replaced it with a step chirp sound that can be done in the background
via quick high notes on the speaker.
This seems to have fixed the timing bug, but like any timing bug,
I'll have to exercise it for like a month before I know.