Modify bridge and jupiter to use unicodes, not just ascii.
This is almost a rewrite - certainly a sweeping change.
Some sections are just commented out for now, like most of tpencode.c.
I don't want to work that hard to rewrite it unless I know we need it,
and I'm not sure we need it any more.
So with that commented out, the -a option doesn't do anything.
There will be many more commits I'm sure;
this is just the start of internationalization.