Let the arrow keys pass through to the running program, and interrupt speech.
This is a big change, a lot of code, for a small benefit,
and it is likely to break something else that I haven't discovered yet.
The timing and race conditions make automatic cursor tracking in screen mode
very fragile.
I hope I don't have to backthis change out some day.
We should be able to keep at least the device driver and bridge enhancements,
it's then a matter of whether and how an adapter might use them.